+++ ila/README.md
@@ -1,0 +1,59 @@
+# OrangeCrab Development Setup
+
+This sub-project contains some demo applications for targeting the [OrangeCrab FPGA r0.2.1](https://orangecrab-fpga.github.io/orangecrab-hardware/docs/r0.2.1) in combination with the OrangeCrab Pmod / Debug Expander board and an [Adafruit FT232H JTAG Debugger](https://learn.adafruit.com/adafruit-ft232h-breakout).
+
+It can to be used as a rapid prototyping template or as a starter for larger project setups.
+
+## Build Architecture
+
+The setup uses a combination of `cabal` and `make`, where
+* `cabal` is used to handle everything Haskell / Clash related up till the point of generating HDL via running the `clash` compiler itself, and
+* `make` is used to run `clash`, RTL synthesis, Place and Route, and for programming the FPGA.
+
+### Cabal
+
+In terms of Haskell / Clash package management, everything is fairly standard. Have a look at the [`orangecrab.cabal`](orangecrab.cabal) and the [`cabal.project`](../cabal.project) files for the particular details. The only special juice is given by the fact that of the additionally used `top` directory, which can contain multiple files to choose among different top entities to be built. They are managed as named libraries within the `orangecrab.cabal` to make them all available via `cabal repl`. Anything else is managed via the `makefile` instead.
+
+### Make
+
+A straightforward [`makefile`](makefile) is used on top of `cabal` to handle any additional calls which are not covered by the cabal build system yet.
+
+#### Configuration
+
+The `makefile` includes the additional configuration file [`build.cfg`](build.cfg), which can be used to link all of the required external build tools. To this end, simply copy the file to `build.cfg.local` on your local machine and modify it towards your personal needs. Any changes made to the copy will be preferred in comparison to original `build.cfg`. Just leave out the overrides, if you like to stick with the defaults instead.
+
+Some references on where to get all of the required external tooling are also given as part of the configuration file.
+
+The particular top entity to be built can be selected via overriding the `NAME` parameter in the `build.cfg.local`. Just set it to the name of the file in the `top` directory shall be built (without the `.hs` ending). See the `makefile` for the pre-configured default.
+
+#### Usage
+
+The `makefile` offers the follwing commands (to be invoked via `make CMD`) where each command depends on the output generated by all of the former ones according to the order of the list below. The `bitstream` command is the `make` default.
+
+| CMD         | Description                                         |
+| ----------- | --------------------------------------------------- |
+| `hdl`       | generates Verilog code from the selected TopEntity  |
+| `synth`     | runs RTL synthesis                                  |
+| `netlist`   | runs Place and Route                                |
+| `bitstream` | generates a bitstream to be transferred to the FPGA |
+| `upload`    | copies the generated bitstream to FPGA              |
+
+The output of each command is stored in a corresponding sub-directory within an automatically created `_build` folder. Use `make clean`, if you like to rebuild everything in the `_build` folder from scratch (`make clean` does **not** run `cabal clean`).
+
+## Demo Applications
+
+### Blink
+
+Simple blinking example using the RGB LED on the OrangeCrab r0.2.1 board.
+
+### Gray
+
+Visualization of a wrapping 8-bit Gray counter via LEDs. A [Digilent Pmod 8LD](https://digilent.com/reference/pmod/pmod8ld/start) must be connected to the `PMOD_3` connector for this. The example utilizes the Gray encoder of [graycode](../graycode).
+
+### Buttons
+
+A simple direct mapping of 4 buttons to 4 LEDs. A [Digilent Pmod 8LD](https://digilent.com/reference/pmod/pmod8ld/start) must be connected to the `PMOD_3` connector and a [Digilent Pmod BTN](https://digilent.com/reference/pmod/pmodbtn/start) to the `PMOD_2` connector for this.
+
+### Uart
+
+Some example of passing serial data at 9600 baud from an external host through the FPGA. A [Digilent Pmod USBUART](https://digilent.com/reference/pmod/pmodbtn/start) must be connected to `PMOD_1` for this. The example utilizes the FIFO of [fifo](../fifo).
