+++ ila/fourmolu.yaml
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2024 Google LLC
+#
+# SPDX-License-Identifier: CC0-1.0
+indentation: 2
+column-limit: 90
