+++ ila/src/Domain.hs
@@ -1,0 +1,24 @@
+{-# LANGUAGE NumericUnderscores #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+{- |
+Module      : Domain
+Copyright   : Copyright © 2024 QBayLogic B.V.
+License     : MIT
+Maintainer  : QBayLogic B.V.
+Stability   : experimental
+Portability : POSIX
+
+OrangeCrab / Lattice ECP5-85F specific clock domains.
+-}
+module Domain where
+
+import Clash.Prelude
+
+-- | 48 MHz oscillator clock of the OrangeCrab board.
+createDomain
+  vSystem
+    { vName = "Dom48"
+    , vResetPolarity = ActiveLow
+    , vPeriod = hzToPeriod 48_000_000
+    }
