--- orangecrab/orangecrab.cabal
@@ -1,113 +1,0 @@
-cabal-version:       2.4
-name:                orangecrab
-synopsis:            orangecrab
-description:         Demo applications for the OrangeCrab r0.2.1
-version:             0.1
-License:             MIT
-license-file:        LICENSE
-author:              QBayLogic B.V.
-maintainer:          devops@qbaylogic.com
-Copyright:           Copyright © 2024 QBayLogic B.V.
-Category:            Hardware
-
-common common-options
-  default-extensions:
-    DataKinds
-    DefaultSignatures
-    DeriveAnyClass
-    DerivingStrategies
-    LambdaCase
-    NoStarIsType
-    TypeFamilies
-    UnicodeSyntax
-    ViewPatterns
-
-    -- TemplateHaskell is used to support convenience functions such as
-    -- 'listToVecTH' and 'bLit'.
-    TemplateHaskell
-    QuasiQuotes
-
-    -- Prelude isn't imported by default as Clash offers Clash.Prelude
-    NoImplicitPrelude
-  ghc-options:
-    -Wall -Wcompat
-    -- Plugins to support type-level constraint solving on naturals
-    -fplugin GHC.TypeLits.Extra.Solver
-    -fplugin GHC.TypeLits.Normalise
-    -fplugin GHC.TypeLits.KnownNat.Solver
-    -fconstraint-solver-iterations=8
-
-    -- Clash needs access to the source code in compiled modules
-    -fexpose-all-unfoldings
-
-    -- Worker wrappers introduce unstable names for functions that might have
-    -- blackboxes attached for them. You can disable this, but be sure to add
-    -- a no-specialize pragma to every function with a blackbox.
-    -fno-worker-wrapper
-
-    -- Strict annotations - while sometimes preventing space leaks - trigger
-    -- optimizations Clash can't deal with. See:
-    --
-    --    https://github.com/clash-lang/clash-compiler/issues/2361
-    --
-    -- These flags disables the optimization.
-    -fno-unbox-small-strict-fields
-    -fno-unbox-strict-fields
-  default-language: GHC2021
-  build-depends:
-    base,
-    hedgehog >= 1.0.3 && < 1.6,
-    split,
-    template-haskell,
-    aeson,
-    th-abstraction,
-    th-expand-syns,
-    hashable,
-    clash-cores,
-    clash-lib,
-    clash-prelude,
-    clash-prelude-hedgehog,
-    clash-protocols,
-    string-interpolate,
-    prettyprinter,
-    mtl,
-    lens,
-    pretty-show,
-    constraints,
-    ghc-typelits-natnormalise,
-    ghc-typelits-extra,
-    ghc-typelits-knownnat
-
-library
-  import: common-options
-  hs-source-dirs: src
-  exposed-modules:
-    SignalFieldSelectors
-    Domain
-    Pmod
-    RGB
-    RingBuffer
-    Ila
-    Communication
-    ConfigGen
-    WishboneUtils
-
-test-suite unittests
-  import: common-options
-  type: exitcode-stdio-1.0
-  hs-source-dirs: tests
-  main-is: Tests.hs
-  other-modules:
-    Tests.RingBuffer
-    Tests.Communication
-  build-depends:
-    orangecrab
-
-library blink
-  import: common-options
-  hs-source-dirs: top
-  build-depends:
-    orangecrab
-  exposed-modules:
-    Blink
-
